Improve layout for large and mobile devices
