+++ app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/view/[climb_uuid]/layout.tsx
@@ -1,0 +1,21 @@
+import { PropsWithChildren } from "react";
+
+import { BoardRouteParametersWithUuid, ParsedBoardRouteParameters } from "@/app/lib/types";
+import { parseBoardRouteParams } from "@/app/lib/util";
+import Col from "antd/es/col";
+import { Content } from "antd/es/layout/layout";
+import Row from "antd/es/row";
+
+interface LayoutProps {
+  params: BoardRouteParametersWithUuid;
+}
+
+export default function ListLayout ({ children, params }: PropsWithChildren<LayoutProps>)  {
+  const parsedParams: ParsedBoardRouteParameters = parseBoardRouteParams(params);
+  return (
+    
+    <Content>
+      {children}
+    </Content>
+  )
+}
